requirements.txt
opencv-python
numpy
torch
pillow
python-escpos
pyserial
ultralytics